Refactor: relocate SearchPanel
