+++ src/components/mainComponents/SearchPanel/searchPanelStyle.css
